requirements.txt
absl-py==2.0.0
accelerate==1.3.0
aider-install==0.1.3
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-cors==0.7.0
aiosignal==1.3.2
airportsdata==20241001
annotated-types==0.6.0
anyio==4.8.0
astor==0.8.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
async-timeout==5.0.1
attrs==23.1.0
av==14.1.0
blake3==1.0.4
blinker==1.9.0
cachetools==5.3.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorful==0.5.6
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
compressed-tensors==0.9.1
contourpy==1.1.1
cupy-cuda12x==13.4.0
cycler==0.12.1
datasets==3.3.1
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1737269734548/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1733236420667/work
decord==0.6.0
deepspeed==0.14.4
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
einops==0.7.0
email_validator==2.2.0
exceptiongroup==1.2.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
fastapi==0.115.8
fastapi-cli==0.0.7
fastrlock==0.8.3
ffmpy==0.5.0
filelock==3.17.0
flash_attn==2.7.4.post1
Flask==3.1.0
fonttools==4.47.0
frozenlist==1.5.0
fsspec==2023.12.2
gguf==0.10.0
google-api-core==2.24.1
google-auth==2.25.2
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.67.0rc1
gradio==5.32.0
gradio_client==1.10.2
groovy==0.1.2
grpcio==1.60.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
imageio==2.37.0
importlib-metadata==7.0.0
importlib-resources==6.1.1
iniconfig==2.0.0
interegular==0.3.3
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1738421264/work
ipywidgets==8.1.5
itsdangerous==2.2.0
jax==0.4.13
jaxlib==0.4.13
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
lark==1.2.2
liger_kernel==0.5.8
llvmlite==0.43.0
lm-format-enforcer==0.10.9
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
mistral_common==1.5.2
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.26.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.1
ninja==1.11.1.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
openai==1.61.1
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
orjson==3.10.18
outcome==1.3.0.post0
outlines==0.1.11
outlines_core==0.1.26
packaging==23.2
pandas==2.2.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
partial-json-parser==0.2.1.1.post5
peft==0.7.1
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==10.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1733232627818/work
plotly==6.0.0
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1737453357274/work
propcache==0.2.1
proto-plus==1.26.0
protobuf==4.25.1
psutil==5.9.7
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==19.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.13.6
pycountry==24.6.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pynndescent==0.5.13
pynvml==12.0.0
pyparsing==3.1.1
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1738270958168/work
qwen-vl-utils==0.0.8
ray==2.40.0
referencing==0.36.2
regex==2023.10.3
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
rsa==4.9
ruff==0.11.12
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
selenium==4.16.0
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sniffio==1.3.0
some-package==0.1
sortedcontainers==2.4.0
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.45.3
sympy==1.13.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.15
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1732615898999/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers @ git+https://github.com/huggingface/transformers@41b9b92b52215bed472c9a534a06abbc3a9a95cd
transformers-stream-generator==0.0.4
trio==0.23.2
trio-websocket==0.11.1
triton==3.1.0
trl==0.17.0
typer==0.15.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
tzdata==2025.1
ujson==5.10.0
umap-learn==0.5.7
urllib3==2.3.0
uv==0.5.30
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.1
vllm==0.7.3
watchfiles==1.0.4
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
websockets==14.2
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
wsproto==1.2.0
xformers==0.0.28.post3
xgrammar==0.1.11
xxhash==3.5.0
yarl==1.18.3
zipp==3.17.0
